Fix some Windows compile issues.
- MSVC complains if an unsigned int is cast to a size_t.
- "and" is not a valid preprocessor logical operator.

BUG=none
R=jamesr@chromium.org

Review URL: https://codereview.chromium.org/829853003

+++ mojo/edk/system/core_test_base.cc
@@ -280,7 +280,7 @@ unsigned CoreTestBase_MockHandleInfo::GetCancelAllAwakablesCallCount() const {
   return cancel_all_awakables_call_count_;
 }
 
-unsigned CoreTestBase_MockHandleInfo::GetAddedAwakableSize() const {
+size_t CoreTestBase_MockHandleInfo::GetAddedAwakableSize() const {
   base::AutoLock locker(lock_);
   return added_awakables_.size();
 }
